Add callback functionality to MidiInterpreter
This is a very basic implementation for now. The constructor takes a
reference to an input stream, and interprets the input as MIDI data.
Clients may register callback functions for the various channel message
types.

Currently, the implementation does not take into account system
exclusive MIDI data; if the stream contains sysex messages, then it will
break, or at least trigger garbage calls to the callback functions.