Clear buffer when changing message type
Previously the buffer would not be cleared when we receive a status
byte. This potentially causes a bug where an earlier data byte would be
used for the new message type for multibyte messages. A well-formed MIDI
stream will not have this behavior, but it's nice to account for it
nonetheless. One possible example sequence would be:

    1. Status byte indicating Note On
    2. Data byte indicating the note. We expect to see a second data
       byte for the velocity after this, so we store this data byte in a
       buffer.
    3. Status byte indicating Control Change. This is unexpected, since
       the previous message is incomplete, but any subsequent bytes
       should part of the CC message.
    4. Data byte indicating the CC number, e.g. 1 for mod wheel.
    5. Data byte indicating the CC value, e.g. 127 for all the way up.

At step #4, the old behavior would be that the interpreter would see the
byte stored in the CharBuffer and trigger the CC handler using the data
byte from step #2. The new behavior clears the buffer at step #3, so
that the CC handler is triggered with the data bytes on #4 and #5.